requirements.txt
black
click~=8.1.7
selenium~=4.19.0
dnspython~=2.6.1